Schedule recurring role downgrading (#5)
+ add functionality for dry-run

+++ active-member-discord-bot/src/main/kotlin/regenerativeag/Discord.kt
@@ -23,11 +23,13 @@ import java.time.LocalDate
 class Discord(
     httpClient: HttpClient,
     private val token: String,
+    private val dryRun: Boolean,
 ) {
     private val logger = KotlinLogging.logger { }
     private val restClient = RestClient(KtorRequestHandler(httpClient, token = token))
     private val usernameCache = UsernameCache(restClient)
     private val channelNameCache = ChannelNameCache(restClient)
+    val roleNameCache = RoleNameCache(restClient)
 
     val postHistory = _PostHistory()
     inner class _PostHistory {
@@ -102,6 +104,7 @@ class Discord(
             val userPairs = userIds.map { it to usernameCache.lookup(it) }
             val sServerId = Snowflake(activeMemberConfig.serverId)
             val sRoleId = Snowflake(roleConfig.roleId)
+            val roleName = roleNameCache.lookup(activeMemberConfig.serverId, roleConfig.roleId)
             val roleIdxByRoleId = activeMemberConfig.roleConfigs.mapIndexed { idx, cfg ->
                 cfg.roleId to idx
             }.toMap()
@@ -111,7 +114,7 @@ class Discord(
                     val sUserId = Snowflake(userId)
                     val userRoles = restClient.guild.getGuildMember(sServerId, sUserId).roles.map { it.value }.toSet()
                     if (roleConfig.roleId in userRoles) {
-                        logger.debug { "$username already has role ${roleConfig.roleId}" }
+                        logger.debug { "$username already has role ${roleConfig.roleId} ($roleName)" }
                     } else {
                         val rolesToRemove = activeMemberConfig.roleConfigs.map { it.roleId }.toSet() - roleConfig.roleId
 
@@ -121,23 +124,29 @@ class Discord(
                                 removeActiveRole(activeMemberConfig.serverId, roleIdToRemove, userId)
                             }
                         }
-                        restClient.guild.addRoleToGuildMember(sServerId, sUserId, sRoleId)
+                        if (rolesToRemove.size > 1) {
+                            logger.warn("Expected at most one role to remove while adding a role to a user... Removing $rolesToRemove from $userId")
+                        }
+
+                        if (dryRun) {
+                            logger.info("Dry run... would have added role=$roleName to $username")
+                        } else {
+                            restClient.guild.addRoleToGuildMember(sServerId, sUserId, sRoleId)
+                            logger.info("Added role $roleName to $username")
+                        }
 
                         val currentRoleLevel = userRoles.mapNotNull { roleIdxByRoleId[it] }.maxOrNull()
                         val newRoleLevel = roleIdxByRoleId[roleConfig.roleId]!!
-                        val shouldWelcome = if (currentRoleLevel == null) {
-                            true
-                        } else {
-                            newRoleLevel > currentRoleLevel
-                        }
-                        if (shouldWelcome) {
+                        val isUpgrade = currentRoleLevel == null || newRoleLevel > currentRoleLevel
+                        if (isUpgrade) {
                             val welcomeConfig = roleConfig.welcomeMessageConfig
-                            restClient.channel.createMessage(Snowflake(welcomeConfig.channel)) {
-                                this.content = welcomeConfig.prefix + "<@$userId>" + welcomeConfig.suffix
-                                val mentions = AllowedMentionsBuilder()
-                                mentions.users.add(Snowflake(userId))
-                                this.allowedMentions = mentions
-                            }
+                            val welcomeMessage = welcomeConfig.createWelcomeMessage(userId)
+                            rooms.postMessage(welcomeMessage, welcomeConfig.channel, listOf(userId))
+                        } else {
+                            val downgradeConfig = activeMemberConfig.downgradeMessageConfig
+                            val previousRoleName = concatRolesToString(activeMemberConfig, rolesToRemove)
+                            val downgradeMessage = downgradeConfig.createDowngradeMessage(username, previousRoleName, roleName)
+                            rooms.postMessage(downgradeMessage, downgradeConfig.channel)
                         }
                     }
                 }
@@ -155,11 +164,16 @@ class Discord(
                 if (userActiveRoles.isEmpty()) {
                     logger.debug { "$username has no active roles to remove" }
                 } else {
-                    userActiveRoles.forEach {
-                        val sRoleId = Snowflake(it)
-                        restClient.guild.deleteRoleFromGuildMember(sServerId, sUserId, sRoleId)
-                        logger.info { "Removed role=$it from $username" }
+                    if (userActiveRoles.size > 1) {
+                        logger.warn("Expected at most one role to remove while removing all active roles from a user... Removing $userActiveRoles from $userId")
                     }
+                    userActiveRoles.forEach { roleId ->
+                       removeActiveRole(activeMemberConfig.serverId, roleId, userId)
+                    }
+                    val removalConfig = activeMemberConfig.removalMessageConfig
+                    val previousRoleName = concatRolesToString(activeMemberConfig, userActiveRoles)
+                    val message = removalConfig.createRemovalMessage(username, previousRoleName)
+                    rooms.postMessage(message, removalConfig.channel)
                 }
             }
         }
@@ -171,28 +185,52 @@ class Discord(
          */
         private fun removeActiveRole(serverId: ServerId, roleId: RoleId, userId: UserId) {
             val username = usernameCache.lookup(userId)
+            val roleName = roleNameCache.lookup(serverId, roleId)
             val sServerId = Snowflake(serverId)
             val sRoleId = Snowflake(roleId)
             runBlocking {
                 val sUserId = Snowflake(userId)
                 val userRoles = restClient.guild.getGuildMember(sServerId, sUserId).roles.map { it.value }
                 if (roleId !in userRoles) {
-                    logger.debug { "$username doesn't have role $roleId to remove" }
+                    logger.debug { "$username doesn't have role $roleId ($roleName) to remove" }
                 } else {
-                    restClient.guild.deleteRoleFromGuildMember(sServerId, sUserId, sRoleId)
+                    if (dryRun) {
+                        logger.info("Dry run... would have removed $roleName from $username")
+                    } else {
+                        restClient.guild.deleteRoleFromGuildMember(sServerId, sUserId, sRoleId)
+                        logger.info("Removed $roleName from $username")
+                    }
                 }
+            }
+        }
+
+        /**
+         * It's possible multiple roles are being removed from the user if there was some manual intervention... or a bug
+         * ...If so, join them together into one string
+         */
+        private fun concatRolesToString(activeMemberConfig: ActiveMemberConfig, roleIds: Set<RoleId>): String {
+            return roleIds.joinToString("+") { removedRoleId ->
+                roleNameCache.lookup(activeMemberConfig.serverId, removedRoleId)
             }
         }
     }
 
-    val server = _Server()
-    inner class _Server {
-        /** Fetch all roles in the server */
-        fun fetchAllRoles(serverId: ServerId): Map<RoleId, String> {
-            val sServerId = Snowflake(serverId)
-            return runBlocking {
-                restClient.guild.getGuildRoles(sServerId).associate {
-                    it.id.value to it.name
+    val rooms = _Rooms()
+    inner class _Rooms() {
+        fun postMessage(message: String, channelId: ChannelId, usersMentioned: List<UserId> = listOf()) {
+            if (dryRun) {
+                val channelName = channelNameCache.lookup(channelId)
+                logger.info { "Dry run... would have posted: \"$message\" in $channelName."}
+            } else {
+                runBlocking {
+                    restClient.channel.createMessage(Snowflake(channelId)) {
+                        this.content = message
+                        this.allowedMentions = AllowedMentionsBuilder().also { builder ->
+                            usersMentioned.forEach { userId ->
+                                builder.users.add(Snowflake(userId))
+                            }
+                        }
+                    }
                 }
             }
         }
@@ -207,11 +245,10 @@ class Discord(
             gateway.events.filterIsInstance<MessageCreate>().onEach { messageCreate ->
                 with(messageCreate.message) {
                     val channelName = channelNameCache.lookup(this.channelId.value)
-                    val threadName = this.thread.value?.name
                     val userId = this.getUserId()
                     val username = usernameCache.lookup(userId)
                     val localDate = this.getLocalDate()
-                    logger.debug { "Message received from $username on $localDate in (channel='$channelName', thread='$threadName')" }
+                    logger.debug { "Message received from $username on $localDate in $channelName" }
                     onMessage(Message(userId, localDate))
                 }
             }.launchIn(gateway)
